requirements.txt
# Q&A Based Interactive Storytelling System
# Requirements File

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# LLM Integration
ollama==0.1.6
openai==1.12.0  # For future API support

# NLP Processing
spacy==3.7.2

# Data Processing
pydantic==2.6.0

# Database
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
